backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pytest>=8.3.0
openai>=1.0.0
python-dotenv>=1.0.0
